Trait rust_cuda::device::BorrowFromRust

pub unsafe trait BorrowFromRust: RustToCuda {
    unsafe fn with_borrow_from_rust<O, F: FnOnce(&Self) -> O>(
        cuda_repr: DeviceBoxConst<<Self as RustToCuda>::CudaRepresentation>, 
        inner: F
    ) -> O;

    unsafe fn with_borrow_from_rust_mut<O, F: FnOnce(&mut Self) -> O>(
        cuda_repr_mut: DeviceBoxMut<<Self as RustToCuda>::CudaRepresentation>, 
        inner: F
    ) -> O;
}

This is supported on non-crate feature host and target_os="cuda" only.

Safety

This is an internal trait and should NEVER be implemented manually

Required methods

unsafe fn with_borrow_from_rust<O, F: FnOnce(&Self) -> O>(
cuda_repr: DeviceBoxConst<<Self as RustToCuda>::CudaRepresentation>,
inner: F
) -> O

Safety

This function is only safe to call iff cuda_repr is the DeviceBoxConst borrowed on the CPU using the corresponding LendToCuda::lend_to_cuda.

unsafe fn with_borrow_from_rust_mut<O, F: FnOnce(&mut Self) -> O>(
cuda_repr_mut: DeviceBoxMut<<Self as RustToCuda>::CudaRepresentation>,
inner: F
) -> O

Safety

This function is only safe to call iff cuda_repr_mut is the DeviceBoxMut borrowed on the CPU using the corresponding LendToCuda::lend_to_cuda_mut. Furthermore, since different GPU threads can access heap storage mutably inside the safe inner scope, there must not be any aliasing between concurrently running threads.