Struct rust_cuda::host::Stream

pub struct Stream<'stream> { /* private fields */ }

Available on crate feature host only.

Implementations

impl<'stream> Stream<'stream>

pub fn with<O>( stream: &mut Stream, inner: impl for<'new_stream> FnOnce(Stream<'new_stream>) -> O ) -> O

Create a new uniquely branded Stream, which can bind async operations to the Stream that they are computed on.

The uniqueness guarantees are provided by using branded types, as inspired by the Ghost Cell paper by Yanovski, J., Dang, H.-H., Jung, R., and Dreyer, D.: https://doi.org/10.1145/3473597.

Examples

The following example shows that two Stream’s with different 'stream lifetime brands cannot be used interchangeably.

use rust_cuda::host::Stream;

fn check_same<'stream>(_stream_a: Stream<'stream>, _stream_b: Stream<'stream>) {}

fn two_streams<'stream_a, 'stream_b>(stream_a: Stream<'stream_a>, stream_b: Stream<'stream_b>) {
    check_same(stream_a, stream_b);
}