Struct rust_cuda::kernel::TypedPtxKernel

pub struct TypedPtxKernel<Kernel> { /* private fields */ }

Available on crate feature host only.

Implementations

impl<Kernel> TypedPtxKernel<Kernel>

pub fn new<T: CompiledKernelPtx<Kernel>>( configure: Option<Box<PtxKernelConfigure>> ) -> Self

impl<Kernel> TypedPtxKernel<Kernel>

pub fn launch0<'kernel, 'stream>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>),

pub fn with0_async<'kernel, 'stream, Ok, Err: From<CudaError>>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>),

pub fn launch0_async<'kernel, 'stream>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>),

pub fn launch1<'kernel, 'stream, A: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A),

pub fn with1_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A),

pub fn launch1_async<'kernel, 'stream, A: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A),

pub fn launch2<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B),

pub fn with2_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B),

pub fn launch2_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B),

pub fn launch3<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C),

pub fn with3_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C),

pub fn launch3_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C),

pub fn launch4<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D),

pub fn with4_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D),

pub fn launch4_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D),

pub fn launch5<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E),

pub fn with5_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>, E::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E),

pub fn launch5_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_>, arg5: E::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E),

pub fn launch6<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F),

pub fn with6_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>, E::AsyncHostType<'stream, '_>, F::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F),

pub fn launch6_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_>, arg5: E::AsyncHostType<'stream, '_>, arg6: F::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F),

pub fn launch7<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G),

pub fn with7_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>, E::AsyncHostType<'stream, '_>, F::AsyncHostType<'stream, '_>, G::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G),

pub fn launch7_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_>, arg5: E::AsyncHostType<'stream, '_>, arg6: F::AsyncHostType<'stream, '_>, arg7: G::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G),

pub fn launch8<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H),

pub fn with8_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>, E::AsyncHostType<'stream, '_>, F::AsyncHostType<'stream, '_>, G::AsyncHostType<'stream, '_>, H::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H),

pub fn launch8_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_>, arg5: E::AsyncHostType<'stream, '_>, arg6: F::AsyncHostType<'stream, '_>, arg7: G::AsyncHostType<'stream, '_>, arg8: H::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H),

pub fn launch9<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, arg9: I::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I),

pub fn with9_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, arg9: I::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>, E::AsyncHostType<'stream, '_>, F::AsyncHostType<'stream, '_>, G::AsyncHostType<'stream, '_>, H::AsyncHostType<'stream, '_>, I::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I),

pub fn launch9_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_>, arg5: E::AsyncHostType<'stream, '_>, arg6: F::AsyncHostType<'stream, '_>, arg7: G::AsyncHostType<'stream, '_>, arg8: H::AsyncHostType<'stream, '_>, arg9: I::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I),

pub fn launch10<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, arg9: I::SyncHostType, arg10: J::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J),

pub fn with10_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, arg9: I::SyncHostType, arg10: J::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>, E::AsyncHostType<'stream, '_>, F::AsyncHostType<'stream, '_>, G::AsyncHostType<'stream, '_>, H::AsyncHostType<'stream, '_>, I::AsyncHostType<'stream, '_>, J::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J),

pub fn launch10_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_>, arg5: E::AsyncHostType<'stream, '_>, arg6: F::AsyncHostType<'stream, '_>, arg7: G::AsyncHostType<'stream, '_>, arg8: H::AsyncHostType<'stream, '_>, arg9: I::AsyncHostType<'stream, '_>, arg10: J::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J),

pub fn launch11<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter, K: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, arg9: I::SyncHostType, arg10: J::SyncHostType, arg11: K::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J, K),

pub fn with11_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter, K: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, arg9: I::SyncHostType, arg10: J::SyncHostType, arg11: K::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>, E::AsyncHostType<'stream, '_>, F::AsyncHostType<'stream, '_>, G::AsyncHostType<'stream, '_>, H::AsyncHostType<'stream, '_>, I::AsyncHostType<'stream, '_>, J::AsyncHostType<'stream, '_>, K::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J, K),

pub fn launch11_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter, K: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_>, arg5: E::AsyncHostType<'stream, '_>, arg6: F::AsyncHostType<'stream, '_>, arg7: G::AsyncHostType<'stream, '_>, arg8: H::AsyncHostType<'stream, '_>, arg9: I::AsyncHostType<'stream, '_>, arg10: J::AsyncHostType<'stream, '_>, arg11: K::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J, K),

pub fn launch12<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter, K: CudaKernelParameter, L: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, arg9: I::SyncHostType, arg10: J::SyncHostType, arg11: K::SyncHostType, arg12: L::SyncHostType ) -> CudaResult<()>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J, K, L),

pub fn with12_async<'kernel, 'stream, Ok, Err: From<CudaError>, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter, K: CudaKernelParameter, L: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::SyncHostType, arg2: B::SyncHostType, arg3: C::SyncHostType, arg4: D::SyncHostType, arg5: E::SyncHostType, arg6: F::SyncHostType, arg7: G::SyncHostType, arg8: H::SyncHostType, arg9: I::SyncHostType, arg10: J::SyncHostType, arg11: K::SyncHostType, arg12: L::SyncHostType, inner: impl FnOnce(&'kernel mut Self, Stream<'stream>, &LaunchConfig, A::AsyncHostType<'stream, '_>, B::AsyncHostType<'stream, '_>, C::AsyncHostType<'stream, '_>, D::AsyncHostType<'stream, '_>, E::AsyncHostType<'stream, '_>, F::AsyncHostType<'stream, '_>, G::AsyncHostType<'stream, '_>, H::AsyncHostType<'stream, '_>, I::AsyncHostType<'stream, '_>, J::AsyncHostType<'stream, '_>, K::AsyncHostType<'stream, '_>, L::AsyncHostType<'stream, '_>) -> Result<Ok, Err> ) -> Result<Ok, Err>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J, K, L),

pub fn launch12_async<'kernel, 'stream, A: CudaKernelParameter, B: CudaKernelParameter, C: CudaKernelParameter, D: CudaKernelParameter, E: CudaKernelParameter, F: CudaKernelParameter, G: CudaKernelParameter, H: CudaKernelParameter, I: CudaKernelParameter, J: CudaKernelParameter, K: CudaKernelParameter, L: CudaKernelParameter>( &'kernel mut self, stream: Stream<'stream>, config: &LaunchConfig, arg1: A::AsyncHostType<'stream, '_>, arg2: B::AsyncHostType<'stream, '_>, arg3: C::AsyncHostType<'stream, '_>, arg4: D::AsyncHostType<'stream, '_>, arg5: E::AsyncHostType<'stream, '_>, arg6: F::AsyncHostType<'stream, '_>, arg7: G::AsyncHostType<'stream, '_>, arg8: H::AsyncHostType<'stream, '_>, arg9: I::AsyncHostType<'stream, '_>, arg10: J::AsyncHostType<'stream, '_>, arg11: K::AsyncHostType<'stream, '_>, arg12: L::AsyncHostType<'stream, '_> ) -> CudaResult<Async<'static, 'stream, (), NoCompletion>>
where Kernel: FnOnce(&mut Launcher<'stream, 'kernel, Kernel>, A, B, C, D, E, F, G, H, I, J, K, L),