Trait rust_cuda::device::BorrowFromRust

pub trait BorrowFromRust: RustToCuda {
    unsafe fn with_borrow_from_rust<O, F: FnOnce(&ShallowCopy<Self>) -> O>(
        cuda_repr: DeviceConstRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>,
        inner: F
    ) -> O;
    unsafe fn with_borrow_from_rust_mut<O, F: FnOnce(&mut ShallowCopy<Self>) -> O>(
        cuda_repr_mut: DeviceMutRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>,
        inner: F
    ) -> O;
    unsafe fn with_moved_from_rust<O, F: FnOnce(Self) -> O>(
        cuda_repr_mut: DeviceMutRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>,
        inner: F
    ) -> O
    where
        Self: Sized + SafeDeviceCopy,
        <Self as RustToCuda>::CudaRepresentation: SafeDeviceCopy;
}

Available on non-crate feature host and target_os="cuda" only.

Required Methods

unsafe fn with_borrow_from_rust<O, F: FnOnce(&ShallowCopy<Self>) -> O>(
cuda_repr: DeviceConstRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>,
inner: F
) -> O

Safety

This function is only safe to call iff cuda_repr is the DeviceConstRef borrowed on the CPU using the corresponding LendToCuda::lend_to_cuda.

unsafe fn with_borrow_from_rust_mut<O, F: FnOnce(&mut ShallowCopy<Self>) -> O>(
cuda_repr_mut: DeviceMutRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>,
inner: F
) -> O

Safety

This function is only safe to call iff cuda_repr_mut is the DeviceMutRef borrowed on the CPU using the corresponding LendToCuda::lend_to_cuda_mut. Furthermore, since different GPU threads can access heap storage mutably inside the safe inner scope, there must not be any aliasing between concurrently running threads.

unsafe fn with_moved_from_rust<O, F: FnOnce(Self) -> O>(
cuda_repr_mut: DeviceMutRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>,
inner: F
) -> Owhere
Self: Sized + SafeDeviceCopy,
<Self as RustToCuda>::CudaRepresentation: SafeDeviceCopy,

Safety

This function is only safe to call iff cuda_repr is the DeviceMutRef borrowed on the CPU using the corresponding LendToCuda::move_to_cuda.