Trait rust_cuda::host::LendToCuda

pub unsafe trait LendToCuda: RustToCuda {
    fn lend_to_cuda<O, F: FnOnce(HostDeviceBoxConst<'_, Self::CudaRepresentation>) -> CudaResult<O>>(
        &self, 
        inner: F
    ) -> CudaResult<O>;

    fn lend_to_cuda_mut<O, F: FnOnce(HostDeviceBoxMut<'_, Self::CudaRepresentation>) -> CudaResult<O>>(
        &mut self, 
        inner: F
    ) -> CudaResult<O>;
}

Safety

This trait should ONLY be derived automatically using #[derive(LendToCuda)]

Required methods

fn lend_to_cuda<O, F: FnOnce(HostDeviceBoxConst<'_, Self::CudaRepresentation>) -> CudaResult<O>>(
&self,
inner: F
) -> CudaResult<O>

Lends an immutable copy of &self to CUDA:

code in the CUDA kernel can only access &self through the DeviceBoxConst inside the closure

after the closure, &self will not have changed

Errors

Returns a rustacuda::errors::CudaError iff an error occurs inside CUDA

fn lend_to_cuda_mut<O, F: FnOnce(HostDeviceBoxMut<'_, Self::CudaRepresentation>) -> CudaResult<O>>(
&mut self,
inner: F
) -> CudaResult<O>

Lends a mutable copy of &mut self to CUDA:

code in the CUDA kernel can only access &mut self through the DeviceBoxMut inside the closure

after the closure, &mut self might have changed in the following ways:

to avoid aliasing, each CUDA thread gets its own shallow copy of &mut self, i.e. any shallow changes will NOT be reflected after the closure

each CUDA thread can access the same heap allocated storage, i.e. any deep changes will be reflected after the closure

Errors

Returns a rustacuda::errors::CudaError iff an error occurs inside CUDA

Loading content...