Trait rust_cuda::host::LendToCuda

pub trait LendToCuda: RustToCuda {
    fn lend_to_cuda<O, E: From<CudaError>, F: FnOnce(HostAndDeviceConstRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>) -> Result<O, E>>(
        &self, 
        inner: F
    ) -> Result<O, E>;

    fn lend_to_cuda_mut<O, E: From<CudaError>, F: FnOnce(HostAndDeviceMutRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>) -> Result<O, E>>(
        &mut self, 
        inner: F
    ) -> Result<O, E>;

    fn move_to_cuda<O, E: From<CudaError>, F: FnOnce(HostAndDeviceOwned<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>) -> Result<O, E>>(
        self, 
        inner: F
    ) -> Result<O, E>
    where
        Self: Sized + SafeDeviceCopy,
        <Self as RustToCuda>::CudaRepresentation: SafeDeviceCopy,
        <Self as RustToCuda>::CudaAllocation: EmptyCudaAlloc;
}

This is supported on crate feature host only.

Required methods

fn lend_to_cuda<O, E: From<CudaError>, F: FnOnce(HostAndDeviceConstRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>) -> Result<O, E>>(
&self,
inner: F
) -> Result<O, E>

Lends an immutable copy of &self to CUDA:

code in the CUDA kernel can only access &self through the DeviceConstRef inside the closure

after the closure, &self will not have changed

Errors

Returns a rustacuda::errors::CudaError iff an error occurs inside CUDA

fn lend_to_cuda_mut<O, E: From<CudaError>, F: FnOnce(HostAndDeviceMutRef<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>) -> Result<O, E>>(
&mut self,
inner: F
) -> Result<O, E>

Lends a mutable copy of &mut self to CUDA:

code in the CUDA kernel can only access &mut self through the DeviceMutRef inside the closure

after the closure, &mut self might have changed in the following ways:

to avoid aliasing, each CUDA thread gets its own shallow copy of &mut self, i.e. any shallow changes will NOT be reflected after the closure

each CUDA thread can access the same heap allocated storage, i.e. any deep changes will be reflected after the closure

Errors

Returns a rustacuda::errors::CudaError iff an error occurs inside CUDA

fn move_to_cuda<O, E: From<CudaError>, F: FnOnce(HostAndDeviceOwned<'_, DeviceAccessible<<Self as RustToCuda>::CudaRepresentation>>) -> Result<O, E>>(
self,
inner: F
) -> Result<O, E> where
Self: Sized + SafeDeviceCopy,
<Self as RustToCuda>::CudaRepresentation: SafeDeviceCopy,
<Self as RustToCuda>::CudaAllocation: EmptyCudaAlloc,

Moves self to CUDA iff self is SafeDeviceCopy

Errors

Returns a rustacuda::errors::CudaError iff an error occurs inside CUDA